src/processors/requirements.txt
confluent-kafka
python-dotenv
apache-flink
slack-sdk